refactor naming and organization
