+++ grid/barriers.py
